+++ delivery-service/src/test/java/com/firstlogistics/deliverservice/application/DeliveryCommandServiceTest.java
@@ -1108,68 +1108,21 @@ class DeliveryCommandServiceTest {
 		}
 	}
 
-	// ===== 배송 취소 테스트 (API) =====
+	// ===== 배송 취소 테스트 =====
 
 	@Nested
-	@DisplayName("배송 취소 실패 (API)")
+	@DisplayName("배송 취소 실패")
 	class CancelDeliveryFail {
-
-		@Test
-		@DisplayName("존재하지 않는 배송")
-		void cancelDelivery_fail_deliveryNotFound() {
-			// given
-			UUID deliveryId = UUID.randomUUID();
-			UUID userId = UUID.randomUUID();
-			ChangeDeliveryStatusCommand command = ChangeDeliveryStatusCommand.of(deliveryId, "MASTER", userId);
-
-			given(deliveryRepository.findById(DeliveryId.of(deliveryId))).willReturn(Optional.empty());
-
-			// when
-			Throwable throwable = catchThrowable(() -> deliveryCommandService.cancelDelivery(command));
-			log.info("throwable = {}", throwable.getMessage());
-
-			// then
-			assertThat(throwable)
-				.isInstanceOf(DeliveryException.class)
-				.hasFieldOrPropertyWithValue("errorCode", DeliveryErrorCode.DELIVERY_NOT_FOUND);
-		}
-
-		@Test
-		@DisplayName("권한 없음 (COMPANY_MANAGER)")
-		void cancelDelivery_fail_accessDenied() {
-			// given
-			UUID deliveryId = UUID.randomUUID();
-			UUID userId = UUID.randomUUID();
-			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.CREATED);
-			ChangeDeliveryStatusCommand command = ChangeDeliveryStatusCommand.of(deliveryId, "COMPANY_MANAGER", userId);
-
-			given(deliveryRepository.findById(DeliveryId.of(deliveryId))).willReturn(Optional.of(delivery));
-			willThrow(new DeliveryException(DeliveryErrorCode.DELIVERY_ACCESS_DENIED))
-				.given(deliveryPermissionValidator).validate(any(), eq("COMPANY_MANAGER"), eq(userId), any());
-
-			// when
-			Throwable throwable = catchThrowable(() -> deliveryCommandService.cancelDelivery(command));
-			log.info("throwable = {}", throwable.getMessage());
-
-			// then
-			assertThat(throwable)
-				.isInstanceOf(DeliveryException.class)
-				.hasFieldOrPropertyWithValue("errorCode", DeliveryErrorCode.DELIVERY_ACCESS_DENIED);
-		}
 
 		@Test
 		@DisplayName("CREATED 아닌 상태 (FOR_HUB_MOVING)")
 		void cancelDelivery_fail_invalidStatus() {
 			// given
 			UUID deliveryId = UUID.randomUUID();
-			UUID userId = UUID.randomUUID();
 			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.FOR_HUB_MOVING);
-			ChangeDeliveryStatusCommand command = ChangeDeliveryStatusCommand.of(deliveryId, "MASTER", userId);
-
-			given(deliveryRepository.findById(DeliveryId.of(deliveryId))).willReturn(Optional.of(delivery));
 
 			// when
-			Throwable throwable = catchThrowable(() -> deliveryCommandService.cancelDelivery(command));
+			Throwable throwable = catchThrowable(() -> deliveryCommandService.cancelDelivery(delivery));
 			log.info("throwable = {}", throwable.getMessage());
 
 			// then
@@ -1180,7 +1133,7 @@ class DeliveryCommandServiceTest {
 	}
 
 	@Nested
-	@DisplayName("배송 취소 성공 (API)")
+	@DisplayName("배송 취소 성공")
 	class CancelDeliverySuccess {
 
 		@Test
@@ -1188,15 +1141,12 @@ class DeliveryCommandServiceTest {
 		void cancelDelivery_success() {
 			// given
 			UUID deliveryId = UUID.randomUUID();
-			UUID userId = UUID.randomUUID();
 			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.CREATED);
-			ChangeDeliveryStatusCommand command = ChangeDeliveryStatusCommand.of(deliveryId, "MASTER", userId);
 
-			given(deliveryRepository.findById(DeliveryId.of(deliveryId))).willReturn(Optional.of(delivery));
 			given(deliveryRepository.save(any(Delivery.class))).willAnswer(inv -> inv.getArgument(0));
 
 			// when
-			ChangeDeliveryStatusResult result = deliveryCommandService.cancelDelivery(command);
+			ChangeDeliveryStatusResult result = deliveryCommandService.cancelDelivery(delivery);
 
 			// then
 			assertThat(result.deliveryId()).isEqualTo(deliveryId);
@@ -1210,15 +1160,12 @@ class DeliveryCommandServiceTest {
 		void cancelDelivery_success_eventPublished() {
 			// given
 			UUID deliveryId = UUID.randomUUID();
-			UUID userId = UUID.randomUUID();
 			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.CREATED);
-			ChangeDeliveryStatusCommand command = ChangeDeliveryStatusCommand.of(deliveryId, "MASTER", userId);
 
-			given(deliveryRepository.findById(DeliveryId.of(deliveryId))).willReturn(Optional.of(delivery));
 			given(deliveryRepository.save(any(Delivery.class))).willAnswer(inv -> inv.getArgument(0));
 
 			// when
-			deliveryCommandService.cancelDelivery(command);
+			deliveryCommandService.cancelDelivery(delivery);
 
 			// then
 			ArgumentCaptor<DeliveryStatusChangedEvent> eventCaptor = ArgumentCaptor.forClass(DeliveryStatusChangedEvent.class);
@@ -1229,48 +1176,23 @@ class DeliveryCommandServiceTest {
 		}
 	}
 
-	// ===== 배송 취소 테스트 (Kafka - cancelByOrder) =====
+	// ===== 시스템 배송 취소 테스트 =====
 
 	@Nested
-	@DisplayName("주문 취소에 의한 배송 취소 실패")
-	class CancelByOrderFail {
-
-		@Test
-		@DisplayName("존재하지 않는 orderId")
-		void cancelByOrder_fail_deliveryNotFound() {
-			// given
-			UUID orderId = UUID.randomUUID();
-
-			given(deliveryRepository.findByOrderId(orderId)).willReturn(Optional.empty());
-
-			// when
-			Throwable throwable = catchThrowable(() -> deliveryCommandService.cancelByOrder(orderId));
-			log.info("throwable = {}", throwable.getMessage());
-
-			// then
-			assertThat(throwable)
-				.isInstanceOf(DeliveryException.class)
-				.hasFieldOrPropertyWithValue("errorCode", DeliveryErrorCode.DELIVERY_NOT_FOUND);
-		}
-	}
-
-	@Nested
-	@DisplayName("주문 취소에 의한 배송 취소 성공")
-	class CancelByOrderSuccess {
+	@DisplayName("시스템 배송 취소 성공")
+	class CancelDeliveryBySystemSuccess {
 
 		@Test
 		@DisplayName("무조건 취소 (어떤 상태든)")
-		void cancelByOrder_success() {
+		void cancelDeliveryBySystem_success() {
 			// given
-			UUID orderId = UUID.randomUUID();
 			UUID deliveryId = UUID.randomUUID();
 			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.FOR_HUB_MOVING);
 
-			given(deliveryRepository.findByOrderId(orderId)).willReturn(Optional.of(delivery));
 			given(deliveryRepository.save(any(Delivery.class))).willAnswer(inv -> inv.getArgument(0));
 
 			// when
-			deliveryCommandService.cancelByOrder(orderId);
+			deliveryCommandService.cancelDeliveryBySystem(delivery);
 
 			// then
 			ArgumentCaptor<Delivery> captor = ArgumentCaptor.forClass(Delivery.class);
@@ -1279,36 +1201,16 @@ class DeliveryCommandServiceTest {
 		}
 
 		@Test
-		@DisplayName("이미 CANCELLED면 멱등 처리 (중복 취소 무시)")
-		void cancelByOrder_success_alreadyCancelled() {
-			// given
-			UUID orderId = UUID.randomUUID();
-			UUID deliveryId = UUID.randomUUID();
-			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.CANCELLED);
-
-			given(deliveryRepository.findByOrderId(orderId)).willReturn(Optional.of(delivery));
-
-			// when
-			deliveryCommandService.cancelByOrder(orderId);
-
-			// then
-			then(deliveryRepository).should(times(0)).save(any(Delivery.class));
-			then(deliveryEventPublisher).should(times(0)).publishDeliveryStatusChanged(any());
-		}
-
-		@Test
 		@DisplayName("취소 시 이벤트 발행")
-		void cancelByOrder_success_eventPublished() {
+		void cancelDeliveryBySystem_success_eventPublished() {
 			// given
-			UUID orderId = UUID.randomUUID();
 			UUID deliveryId = UUID.randomUUID();
 			Delivery delivery = stubDeliveryWithRoutes(deliveryId, DeliveryStatus.HUB_WAITING);
 
-			given(deliveryRepository.findByOrderId(orderId)).willReturn(Optional.of(delivery));
 			given(deliveryRepository.save(any(Delivery.class))).willAnswer(inv -> inv.getArgument(0));
 
 			// when
-			deliveryCommandService.cancelByOrder(orderId);
+			deliveryCommandService.cancelDeliveryBySystem(delivery);
 
 			// then
 			ArgumentCaptor<DeliveryStatusChangedEvent> eventCaptor = ArgumentCaptor.forClass(DeliveryStatusChangedEvent.class);
